requirements.txt
streamlit==1.27.2
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.98.0
httplib2==0.22.0
sqlalchemy==2.0.20
